requirements.txt
torch
numpy
tqdm
scikit-learn
setuptools
transformers
langchain
langchain-community
langchain-ollama
pandas
openai
anthropic
google-generativeai
PyYAML
